+++ models.py
@@ -1,8 +1,6 @@
 import timm
 import torch
 import torch.nn as nn
-import torchvision.models as models
-from typing import Tuple